src/test/java/com/cakefactory/controllers/OrderControllerTest.java
package com.cakefactory.controllers;

import com.cakefactory.service.OrderService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.BDDMockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.BDDMockito.willDoNothing;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;

@WebMvcTest(OrderController.class)
@AutoConfigureMockMvc
class OrderControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    OrderService orderService;

    @BeforeEach
    public void init(){
        willDoNothing().given(orderService).placeOrder();
    }

    @Test
    public void testPlaceOrderScreen() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.post("/order"))
                .andExpect(MockMvcResultMatchers.status().is2xxSuccessful());
        verify(orderService, times(1)).placeOrder();
    }
}